chore: update pest to 2.x & allow for comments within enums & structs

+++ omg-idl-gen/files/test-vectors/include_directive/expected.rs
@@ -5,7 +5,7 @@ pub mod ModuleA {
 
     #[allow(dead_code)]
     #[allow(non_camel_case_types)]
-    pub type dim1 = [i32;2];
+    pub type dim1 = [i32;2 as usize];
 
     #[allow(dead_code)]
     #[allow(non_camel_case_types)]
